requirements.txt
flask
flask-restx
flask-httpauth
openai>=1.12.0
pillow
pytest
werkzeug
pyathena>=2.0.0
langchain>=0.0.350
pandas>=2.0.0
sqlalchemy>=2.0.0
langchain-community
python-dotenv>=1.0.0
langchain-openai
anthropic
langchain-anthropic
flask-login
flask-sqlalchemy
requests
oauthlib
pyOpenSSL